added the dvc_data_pipeline and shell script to create the whole pipeline and also tear down the whole pipeline in one run
