--- dvc_data_pipeline/ak.sh
@@ -1,16 +1,0 @@
-#!/bin/bash
-
-set -e  # Exit on error
-
-# Load configuration from config.env
-CONFIG_FILE=".env"
-
-if [ -f "$CONFIG_FILE" ]; then
-    echo "Loading configuration from $CONFIG_FILE..."
-    source "$CONFIG_FILE"
-else
-    echo "Error: Configuration file $CONFIG_FILE not found!"
-    exit 1
-fi
-
-aws ecs update-service --cluster "$ECS_CLUSTER_NAME" --service "$ECS_SERVICE_NAME" --desired-count 1
