+++ dvc_data_pipeline/example.env
@@ -1,0 +1,16 @@
+AWS_ACCESS_KEY_ID='aws access key'
+AWS_SECRET_ACCESS_KEY='aws secret access key'
+AWS_REGION='aws account region'
+ROLE_NAME='role name'
+POLICY_NAME=aka_DataPipelinePolicy
+LOG_GROUP_NAME=/aws/mlops-pipeline_aka
+BUCKET_NAME=mlops-data-pipeline-aka
+ECS_CLUSTER_NAME=MLOpsCluster_aka
+ECS_TASK_DEFINITION=MLOpsTask_aka
+ECS_SERVICE_NAME=MLOpsTask_aka_contianer
+ECR_REPOSITORY_NAME=mlops-data-processing_aka
+STEP_FUNCTIONS_NAME=MLOpsStepFunction_aka
+LAMBDA_FUNCTION_NAME=MLOpsTrigger_aka
+CODEPIPELINE_NAME=MLOpsCodePipeline_aka
+CODEBUILD_PROJECT_NAME=MLOpsCodeBuild_aka
+AWS_ACCOUNT_ID='aws account id'
